save enhancements
