+++ src/web/client/common/authenticationProvider.ts
@@ -4,8 +4,9 @@
  */
 
 import * as vscode from 'vscode';
-import { PROVIDER_ID, SCOPE_OPTION, SCOPE_VERB } from './constants';
+import { pathParamToSchema, PROVIDER_ID, SCOPE_OPTION, SCOPE_VERB } from './constants';
 import { ERRORS } from './errorHandler';
+import { dataSourcePropertiesMap } from './localStore';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 export function getHeader(accessToken: string) {
@@ -27,4 +28,17 @@ export async function dataverseAuthentication(dataverseOrgURL: string): Promise<
         vscode.window.showErrorMessage(ERRORS.AUTHORIZATION_FAILED);
     }
     return accessToken;
+}
+
+export function getRequestURLSingleEntity(dataverseOrgUrl: string, entity: string, entityId: string, urlquery: string, entitiesSchemaMap: any, method: string): string {
+    const parameterizedUrl = dataSourcePropertiesMap.get(urlquery) as string;
+    let requestUrl = parameterizedUrl.replace('{dataverseOrgUrl}', dataverseOrgUrl).replace('{entity}', entity).replace('{entityId}', entityId).replace('{api}', dataSourcePropertiesMap.get('api')).replace('{data}', dataSourcePropertiesMap.get('data')).replace('{version}', dataSourcePropertiesMap.get('version'));
+    switch (method) {
+        case 'GET':
+            requestUrl = requestUrl + entitiesSchemaMap.get(pathParamToSchema.get(entity)).get('_query');
+            break;
+        default:
+            break;
+    }
+    return requestUrl;
 }
