+++ src/web/client/common/remoteSaveProvider.ts
@@ -6,9 +6,23 @@
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import * as vscode from 'vscode';
-import { getHeader } from './authenticationProvider';
+import { getHeader, getRequestURLSingleEntity } from './authenticationProvider';
+import { CHARSET, SINGLE_ENTITY_URL_KEY } from './constants';
 import { ERRORS, showErrorDialog } from './errorHandler';
+import { PortalsFS } from './fileSystemProvider';
+import { entitiesSchemaMap } from './localStore';
 import { SaveEntityDetails } from './portalSchemaInterface';
+
+export function registerSaveProvider(accessToken: string, portalsFS: PortalsFS, dataVerseOrgUrl: string, saveDataMap: Map<string, SaveEntityDetails>)
+{
+    vscode.workspace.onDidSaveTextDocument(async (e) => {
+        vscode.window.showInformationMessage('saving file: ' + e.uri);
+        const newFileData = portalsFS.readFile(e.uri);
+        const patchRequestUrl = getRequestURLSingleEntity(dataVerseOrgUrl, saveDataMap.get(e.uri.fsPath)?.getEntityName as string, saveDataMap.get(e.uri.fsPath)?.getEntityId as string, SINGLE_ENTITY_URL_KEY, entitiesSchemaMap, 'PATCH');
+        vscode.window.showInformationMessage(patchRequestUrl)
+        await saveData(accessToken, patchRequestUrl, e.uri, saveDataMap, new TextDecoder(CHARSET).decode(newFileData));
+    });
+}
 
 export async function saveData(accessToken: string, requestUrl: string, fileUri: vscode.Uri, saveDataMap: Map<string, SaveEntityDetails>, value: string) {
     let requestBody = '';
